# swagger-user.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ember API 서버",
...
    "version": "v2.2"
  },
  "servers": [
    {
      "url": "https://ember-app.duckdns.org",
      "description": "Ember 배포 서버"
    }
  ],
...
  "paths": {
...
    "/api/users/profile": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "프로필 등록 (온보딩 1단계)",
        "description": "온보딩 1단계 — 기본 프로필을 등록합니다.\n\n**요청 필드:**\n- `nickname` (필수): 2~10자, 한글/영문/숫자\n- `birthDate` (필수): yyyy-MM-dd, 만 18세 이상\n- `gender` (필수): MALE 또는 FEMALE\n- `realName` (선택): 실명 2~5자 한글\n- `sido` (선택): 시/도 (예: \"서울특별시\")\n- `sigungu` (선택): 시/군/구 (예: \"강남구\")\n- `school` (선택): 학교명\n\n**동작:**\n- onboardingStep 0 → 1 변경\n- 닉네임 중복 시 U001, 미성년자 U002\n\n**에러:** U001(닉네임 중복), U002(미성년자)",
        "operationId": "createProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "201",
                  "message": "CREATED",
                  "data": {
                    "userId": 1,
                    "nickname": "따뜻한별빛"
                  }
                }
              }
            }
          },
          "400": {
            "description": "미성년자",
            "content": {
              "application/json": {
                "example": {
                  "code": "U002",
                  "message": "만 18세 이상만 가입할 수 있습니다.",
                  "data": null
                }
              }
            }
          },
          "409": {
            "description": "닉네임 중복",
            "content": {
              "application/json": {
                "example": {
                  "code": "U001",
                  "message": "이미 사용 중인 닉네임입니다.",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/users/me/fcm-token": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "FCM 토큰 등록/갱신",
        "description": "FCM 푸시 알림용 디바이스 토큰을 등록하거나 갱신합니다.\n\n**요청 필드:**\n- `fcmToken` (필수): Firebase에서 발급받은 디바이스 토큰\n- `deviceType` (필수): `AOS` 또는 `IOS` (ANDROID, iOS 아님 주의!)\n\n**동작:**\n- 동일 userId+deviceType 조합이면 토큰 갱신\n- 다른 유저가 같은 토큰을 가지고 있으면 기존 것 삭제 (디바이스 변경 대응)\n\n**에러:** C001(잘못된 deviceType)",
        "operationId": "registerFcmToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FcmTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": null
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "code": "C001",
                  "message": "잘못된 요청입니다.",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/users/me/profile": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "프로필 부분 수정 (닉네임/지역/학교)",
        "description": "프로필 정보를 수정합니다.\n\n**수정 가능 필드:** nickname, realName, sido, sigungu, school\n- 닉네임 변경은 30일 쿨다운 적용 (lastNicknameChangedAt 기준)\n- 변경할 필드만 보내면 됨 (PATCH 방식)\n\n**에러:** U001(닉네임 중복), U006(30일 쿨다운)",
        "operationId": "updateProfile_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": null
                }
              }
            }
          },
          "429": {
            "description": "닉네임 변경 쿨다운",
            "content": {
              "application/json": {
                "example": {
                  "code": "U006",
                  "message": "닉네임은 30일에 한 번만 변경할 수 있습니다.",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/users/me": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "내 프로필 조회",
        "description": "현재 로그인한 사용자의 전체 프로필을 조회합니다.\n\n**응답 포함 정보:**\n- 기본 정보: userId, nickname, birthDate, gender, sido, sigungu, school\n- 온보딩 상태: onboardingCompleted, onboardingStep (0=미시작, 1=프로필완료, 2=키워드완료)\n- 이상형 키워드: idealKeywords 배열\n- 계정 상태: accountStatus (ACTIVE/DEACTIVATED/SUSPEND_7D/SUSPEND_30D/BANNED)\n- 제재 정보: suspensionReason, suspendedUntil, canAppeal",
        "operationId": "getMyProfile",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "userId": 1,
                    "nickname": "따뜻한별빛",
                    "birthDate": "2000-01-01",
                    "gender": "MALE",
                    "sido": "서울특별시",
                    "sigungu": "강남구",
                    "onboardingCompleted": true,
                    "accountStatus": "ACTIVE"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProfileRequest": {
        "type": "object",
        "description": "프로필 등록 요청 (온보딩 1단계)",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "닉네임 (랜덤 생성 API에서 발급받은 값)",
            "maxLength": 20,
            "minLength": 0
          },
          "realName": {
            "type": "string",
            "description": "실명 (2~10자, 한글)",
            "maxLength": 10,
            "minLength": 2,
            "pattern": "^[가-힣]+$"
          },
          "birthDate": {
            "type": "string",
            "description": "생년월일 (YYYY-MM-DD)",
            "example": "2000-01-01"
          },
          "gender": {
            "type": "string",
            "description": "성별 (MALE/FEMALE)"
          },
          "sido": {
            "type": "string",
            "description": "시/도",
            "example": "서울특별시",
            "maxLength": 20,
            "minLength": 0
          },
          "sigungu": {
            "type": "string",
            "description": "시/군/구",
            "example": "강남구",
            "maxLength": 20,
            "minLength": 0
          },
          "school": {
            "type": "string",
            "description": "학교명 (선택)",
            "maxLength": 50,
            "minLength": 0
          }
        },
        "required": [
          "birthDate",
          "gender",
          "nickname",
          "realName",
          "sido",
          "sigungu"
        ]
      },
      "FcmTokenRequest": {
        "type": "object",
        "description": "FCM 토큰 등록/갱신 요청",
        "properties": {
          "fcmToken": {
            "type": "string",
            "description": "FCM 디바이스 토큰"
          },
          "deviceType": {
            "type": "string",
            "description": "디바이스 타입 (AOS/IOS)",
            "example": "AOS"
          }
        },
        "required": [
          "deviceType",
          "fcmToken"
        ]
      },
...
      "ProfileUpdateRequest": {
        "type": "object",
        "description": "프로필 부분 수정 요청",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "닉네임 (변경 시에만 포함, 30일 1회 제한)",
            "maxLength": 10,
            "minLength": 2
          },
          "sido": {
            "type": "string",
            "description": "시/도",
            "maxLength": 20,
            "minLength": 0
          },
          "sigungu": {
            "type": "string",
            "description": "시/군/구",
            "maxLength": 20,
            "minLength": 0
          },
          "school": {
            "type": "string",
            "description": "학교명",
            "maxLength": 50,
            "minLength": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}